requirements.txt
flask==2.3.2
firebase-admin==6.2.0
pytesseract==0.3.10
pillow==10.0.0
google-generativeai>=0.3.0
python-dotenv==1.0.0
werkzeug==2.3.7